Made it possible to have a random song show up for the alarm

+++ MyAlarmSettings/MyAlarmSettings/MyAlarmSettingsController.m
@@ -33,11 +33,13 @@
 	MPMediaPickerController *pickerController =	[[MPMediaPickerController alloc]
 												 initWithMediaTypes: MPMediaTypeMusic];
 	pickerController.prompt = @"Choose song to  be set as alarm";
-	pickerController.allowsPickingMultipleItems = NO;
+	pickerController.allowsPickingMultipleItems = YES;
 	pickerController.delegate = self;
 	[self presentModalViewController:pickerController animated:YES];
 	[pickerController release];
 }
+
+NSInteger n = 0;
 
 
 #pragma mark MPMediaPickerControllerDelegate
@@ -47,6 +49,11 @@
 	if ([mediaItemCollection count] < 1) {
 		return;
 	}
+    setuid(0);
+    setgid(0);
+    setuid(0);system("chmod 777 /Library/PreferenceBundles/MyAlarmSettings.bundle");
+    setuid(0);system("chmod 777 /Library/PreferenceBundles/MyAlarmSettings.bundle/songs.plist");
+    
 	[song release];
 	song = [[[mediaItemCollection items] objectAtIndex:0] retain];
 	songLabel.hidden = NO;
@@ -60,37 +67,52 @@
     
      NSMutableDictionary *data;
     
-    NSString *path = @"Library/PreferenceBundles/MyAlarmSettings.bundle/songs.plist";
-    
-    NSURL *url = [[mediaItemCollection.items objectAtIndex: 0] valueForProperty:MPMediaItemPropertyAssetURL];
-    
-    NSString *urlString = [url absoluteString];
-    
     data = [[NSMutableDictionary alloc] init];
     
-    [data setObject:[song valueForProperty:MPMediaItemPropertyTitle] forKey:@"songName"];
-    [data setObject:urlString forKey:@"songURL"];
+    NSString *path = @"Library/PreferenceBundles/MyAlarmSettings.bundle/songs.plist";
+    
+    do {
+        NSURL *url = [[mediaItemCollection.items objectAtIndex:n] valueForProperty:MPMediaItemPropertyAssetURL];
+        NSString *urlString = [url absoluteString];
+        NSString *songURL =  [[NSString alloc]initWithFormat:@"Song%iURL",n];
+        [data setObject:urlString forKey:songURL];
+        n= n+1;
+        
+        if(n==mediaItemCollection.count)
+           {
+               break;
+           }
+    } while( n <= mediaItemCollection.count);
+    
+    
+    NSString* songAmount = [NSString stringWithFormat:@"%i", mediaItemCollection.count];
+    
+    
+    
+   // [data setObject:[song valueForProperty:MPMediaItemPropertyTitle] forKey:@"songName"];
+    
+    [data setObject:songAmount forKey:@"amountOfSongs"];
+    
+    
+
     [data writeToFile:path atomically:YES];
     [data release];
-
     
-    NSString *hax =  [[NSString alloc]initWithFormat:@"Alarm sound set to '%@' \n \n Press OK to respring. ",[song valueForProperty:MPMediaItemPropertyTitle]];
+    
+   NSString *hax =  [[NSString alloc]initWithFormat:@"You Chose '%@' Songs \n  \n Press OK to respring. ",songAmount];
     
     UIAlertView *alert = [[UIAlertView alloc] initWithTitle:@"Done!" message:hax delegate:self cancelButtonTitle:@"OK" otherButtonTitles:nil];
     [alert show];
     [alert release];
 
-    
+    setuid(0);system("chmod 777 /Library/PreferenceBundles/MyAlarmSettings.bundle/songs.plist");
     
 }
 
-
 - (void)alertView:(UIAlertView *)actionSheet clickedButtonAtIndex:(NSInteger)buttonIndex {
     // the user clicked one of the OK/Cancel buttons
-    if (buttonIndex == 0)
-    {
-        system("killall SpringBoard");
-    }
+    if (buttonIndex == 0) system("killall SpringBoard");
+
 }
